♻️ Move ob to central TS file
