+++ src/ts/extension/ob.ts
@@ -1,0 +1,2 @@
+const ob = require("../../../out/cljs/owlbear");
+export default ob;
